% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bindings.R
\docType{data}
\name{cim10fr2019}
\alias{cim10fr2019}
\title{Synonym for \code{\link{icd10fr2019}}, with column names in the French language}
\usage{
cim10fr2019()
}
\description{
Synonym for \code{\link{icd10fr2019}}, with column names in the French language
}
\keyword{datasets}
