% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_icd10cm.R
\name{.dl_icd10cm_all}
\alias{.dl_icd10cm_all}
\title{Fetch ICD-10-CM data from the CMS web site}
\usage{
.dl_icd10cm_all(verbose = FALSE, ...)
}
\description{
YEAR-ICD10-Code-Descriptions has flat files, YEAR-ICD10-Code-Tables-Index has
XML
}
\keyword{internal}
