% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/who.R
\name{.dl_icd10who2016}
\alias{.dl_icd10who2016}
\title{Fetch English or French WHO data from online source}
\usage{
.dl_icd10who2016(save_data = TRUE, ...)
}
\arguments{
\item{save_data}{Logical, defaults to `TRUE`}

\item{...}{Arguments passed to internal functions}
}
\description{
This will download the latest ICD-10 codes and descriptions from the WHO, and
save the results in a directory given by `icd.data::get_resource_dir()`.
This defaults to a subdirectory `.icd.data` of the home directory. This is
necessary because it is not permitted to write data back to the installed
package location (and this may not be allowed on a multi-user system,
anyway).
}
\keyword{internal}
