% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R
\name{.exists}
\alias{.exists}
\alias{.get}
\alias{.assign}
\alias{.ls}
\title{Check or get data from environment, not file cache}
\usage{
.exists(var_name)

.get(var_name)

.assign(var_name, value)

.ls()
}
\description{
Check or get data from environment, not file cache
}
\seealso{
\code{\link{.get_getter_name}}
}
\keyword{internal}
