% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R
\name{.fetch}
\alias{.fetch}
\alias{.available}
\title{Gets data, from env cache, file cache, the downloading and parsing if
necessary and possible.}
\usage{
.fetch(var_name, must_work = TRUE, verbose = TRUE, ...)

.available(var_name, verbose = TRUE, ...)
}
\arguments{
\item{var_name}{Character, e.g., \code{"icd10cm2015"}}

\item{must_work}{Single logical}

\item{verbose}{single logical value, defaults to \code{FALSE} in most functions.}

\item{...}{E.g., \code{dx = FALSE} or \code{offline = TRUE}}
}
\description{
Fetch functions are of the form \code{.fetch_icd10cm2015}. This is generic
for all datasets, and calls the data specific \code{parse} function, which
itself will call (if necessary) the data-specific \code{dl} download
function.
}
\keyword{internal}
