% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/who.R
\name{.fetch_icd10_who}
\alias{.fetch_icd10_who}
\title{Use public interface to fetch ICD-10 WHO version}
\usage{
.fetch_icd10_who(concept_id = NULL, year = 2016, lang = "en",
  verbose = FALSE, hier_code = character(), hier_desc = character(),
  ...)
}
\arguments{
\item{concept_id}{This is the id for the code or code group, e.g. "XI"
(Chapter 6), "T90-T98" (A sub-chapter), "E01" (A sub-sub-chapter). You
cannot query a single code with this interface.}

\item{year}{integer 4-digit year}

\item{lang}{Currently it seems only 'en' works}

\item{verbose}{logical}

\item{...}{further arguments passed to self recursively, or \code{.fetch_who_api}}
}
\description{
The user may call this function to install the full WHO ICD-10 definition on
their machine, after which it will be available to \code{icd}. TODO: determine the
best place to save this data.
}
\keyword{internal}
