% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/who.R
\name{.fetch_who_api}
\alias{.fetch_who_api}
\title{Functions to get the WHO ICD-10 2016 data}
\usage{
.fetch_who_api(resource, edition = "icd10", year = 2016, lang = "en",
  verbose = FALSE)
}
\arguments{
\item{resource}{Fragment of URL with specific ICD-10 resource requested}

\item{edition}{icd10}

\item{year}{Four-digit year as integer or character}

\item{lang}{Language string, e.g., "en", "fr", "nl" or others that may be supported in the future.}

\item{verbose}{single logical value, defaults to \code{FALSE} in most functions.}
}
\value{
\code{.fetch_who_api} returns the JSON data, or fails with NULL
}
\description{
This is likely to be not exported in the future, as it is designed for the
transition to icd.data version 1.1 . The user may use the active binding
\code{icd10who2016} as if it is a variable. In some situations, it may be
preferable to call this function. E.g., using the active binding when the
cache directory has not been populated may produce messages. Auto-complete in
Rstudio is unfortunately considered still to be interactive, so these
messages may appear prematurely.
}
\details{
This may be needed (in addition to icd.data::icd10who2016 active binding),
because lazy data and apparently active bindings are not available until the
package is on the search path. This is difficult or impossible to do in a
CRAN compatible way in the 'icd' package, so this function is just for that
purpose, since we can check whether this function exists and call it without
using ::, whereas this is not possible with lazy data or active bindings?
}
\keyword{datasets}
\keyword{internal}
