% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{.get_chapters_fr}
\alias{.get_chapters_fr}
\title{Use the WHO ICD-10 French 'hierarchy' flat file to infer chapter ranges}
\usage{
.get_chapters_fr(save_data = FALSE)
}
\description{
These are UTF-8 encoded. If there are no UTF-8 characters, it seems that R
forces the 'unknown' encoding label. This could be scraped from the web site
directly, which is what \code{.fetch_who_api_chapters()} does, but this is
at least as good.
}
\keyword{internal}
