% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R
\name{.get_fetcher_name}
\alias{.get_fetcher_name}
\alias{.get_fetcher_fun}
\title{Get name or function for fetching a specific ICD data set}
\usage{
.get_fetcher_name(var_name)

.get_fetcher_fun(var_name)
}
\description{
Get name or function for fetching a specific ICD data set
}
\seealso{
\code{\link{.fetch}}
}
\keyword{internal}
