% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd-cm-ver.R
\name{.get_icd10cm_latest}
\alias{.get_icd10cm_latest}
\title{Get the data for the latest ICD-10-CM version in this package.}
\usage{
.get_icd10cm_latest()
}
\description{
Usually, the `icd.data` package should be attached, and therefore appear on
the search list, using `library(icd.data)` or `require(icd.data)`. If it is
not attached, the active binding `icd10cm_latest` cannot find the packages
own data! This function is a possible work-around to get the data without
having to attach the package.
}
\examples{
# if icd.data not attached:
a <- icd.data::icd10cm_latest
b <- icd.data:::.get_icd10cm_latest()
stopifnot(identical(a, b))
# Preferred:
\dontrun{
library(icd.data)
head(icd10cm_latest)
}
}
\keyword{datasets}
\keyword{internal}
