% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd-cm-ver.R
\name{.get_icd10cm_ver}
\alias{.get_icd10cm_ver}
\title{Internal function used to search and maybe prompt when active binding used.}
\usage{
.get_icd10cm_ver(ver, dx, must_work = FALSE, interact = interactive())
}
\arguments{
\item{interact}{Control whether functions thinks it is in interactive mode,
for testing.}
}
\description{
Tries to get from the local environment first, then from resource directory,
and failing that, if interactive, prompts user to download and parse.
}
\keyword{internal}
