% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R
\name{.get_parser_name}
\alias{.get_parser_name}
\title{Parse ICD data (downloading data if needed)}
\usage{
.get_parser_name(var_name)
}
\description{
Parse ICD data (downloading data if needed)
}
\seealso{
\code{\link{.get}}, \code{\link{.get_getter_name}} and \code{\link{.get_parse_icd10cm_name}}
}
