% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-file.R
\name{.get_raw_data_path}
\alias{.get_raw_data_path}
\title{Get the raw data directory}
\usage{
.get_raw_data_path(file_name)
}
\description{
Following Hadley Wickham recommendations in R Packages, this should be in
\code{inst/extdata}. \pkg{devtools} overrides \code{system.file}.
}
\keyword{internal}
