% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R
\name{.ls_icd_data}
\alias{.ls_icd_data}
\title{List the actual data in this package, not bindings}
\usage{
.ls_icd_data()
}
\description{
List the actual data in this package, not bindings
}
\examples{
icd.data:::.ls_icd_data()
}
\keyword{datasets}
\keyword{internal}
