% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cim-10-fr.R
\name{.parse_cim_fr}
\alias{.parse_cim_fr}
\title{Read the definitions of the French edition of ICD-10}
\usage{
.parse_cim_fr(save_data = FALSE, ...)
}
\arguments{
\item{save_data}{single logical value, if \code{TRUE} will save the generated
data}
}
\description{
The short descriptions are capitalized, with accented characters, so leaving as is.
}
\keyword{internal}
