% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-icd10.R
\name{.parse_icd10cm_all}
\alias{.parse_icd10cm_all}
\title{Get all ICD-10-CM codes}
\source{
\url{https://www.cms.gov/Medicare/Coding/ICD10} also available from
  \url{http://www.cdc.gov/nchs/data/icd/icd10cm/2016/ICD10CM_FY2016_code_descriptions.zip}.
}
\usage{
.parse_icd10cm_all(save_data = FALSE,
  offline = getOption("icd.data.offline"), verbose = .verbose(),
  twentysixteen = FALSE, ...)
}
\description{
Gets all ICD-10-CM codes from an archive on the CDC web site.
}
\details{
The factor generation uses \code{sort.default} which is locale dependent.
This meant a lot of time debugging a problem when white space was ignored for
sorting on some platforms, but not others.
}
\references{
\href{https://www.cms.gov/Medicare/Coding/ICD10/downloads/icd-10quickrefer.pdf}{CMS ICD-10 Quick Reference}
  \href{https://www.cdc.gov/nchs/icd/icd10cm.htm#FY\%202019\%20release\%20of\%20ICD-10-CM}{CDC copy of ICD-10-CM for 2019}
}
\keyword{datagen}
\keyword{internal}
