% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cim-10-fr.R
\name{.parse_icd10fr2019}
\alias{.parse_icd10fr2019}
\title{Read the definitions of the French edition of ICD-10}
\usage{
.parse_icd10fr2019(save_data = TRUE, ...)
}
\arguments{
\item{save_data}{single logical value, if \code{TRUE} will save the generated
data}
}
\description{
The short descriptions are capitalized, with accented characters, so leaving
as is.
}
\keyword{internal}
