% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-rtf.R
\name{.rtf_parse_year}
\alias{.rtf_parse_year}
\title{parse RTF description of entire ICD-9-CM for a specific year}
\source{
\url{https://www.cdc.gov/nchs/icd/icd9cm.htm} Navigate to
'Dtab12.zip' in the 2011 data. and similar files run from 1996 to 2011.
}
\usage{
.rtf_parse_year(year = "2011", ..., save_data = FALSE,
  verbose = FALSE, offline = getOption("icd.data.offline"))
}
\arguments{
\item{year}{from 1996 to 2012 (this is what CDC has published). Only 2012
implemented thus far}

\item{save_data}{single logical value, if \code{TRUE} will save the generated
data}

\item{verbose}{single logical value, defaults to \code{FALSE} in most functions.}

\item{offline}{single logical, if \code{TRUE} then don't pull the file from
internet, only return path and file name if the file already exists in the
raw data directory. This is helpful for testing without using the internet.
Setting the option \code{icd.data.offline} will set this globally. For anyone
with an unmetered internet connection, setting this to \code{TRUE} is a good
idea, and the default. It may be set to \code{FALSE}, which is useful for
automated testing when downloads are not desirable. N.b., most people do
not need to download or parse any raw ICD data themselves, so really this
is just for developers.}
}
\description{
Currently only the most recent update is implemented. Note that CMS have
published additional ICD-9-CM billable code lists since the last one from the
CDC: I think these have been the same every year since 2011, though. The last
CDC release is \code{Dtab12.rtf} from 2011.
}
\details{
The file itself is 7 bit ASCII, but has its own internal encoding using
'CP1252.' Test 'Meniere's' disease with lines 24821 to 24822 from 2012 RTF
}
\keyword{datagen}
\keyword{internal}
