% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-file.R
\name{.save_in_data_dir}
\alias{.save_in_data_dir}
\title{Save given variable in package data directory}
\usage{
.save_in_data_dir(var_name, suffix = "", compress = "gzip",
  envir = parent.frame())
}
\arguments{
\item{var_name}{character name of variable or its symbol either of which
would find \code{myvar} in the parent environment, and save it as
\code{myvar.RData} in \code{package_root/data}.}

\item{suffix}{character scalar}

\item{envir}{environment in which to look for the variable to save}

\item{data_path}{path to data directory, default is data in current
directory.}

\item{package_dir}{character containing the directory root of the package
tree in which to save the data. Default is the current working directory.}
}
\value{
invisibly returns the data
}
\description{
File is named \code{varname.suffix.rda}
}
\keyword{internal}
