% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/who.R
\name{.who_api_chapter_names}
\alias{.who_api_chapter_names}
\title{Use WHO API to discover chapters}
\usage{
.who_api_chapter_names(ver = "icd10", year = 2016, lang = "en",
  verbose = .verbose())
}
\description{
Of note, the `WHO` package does not provide access to classifications, just
WHO summary data.
}
\keyword{internal}
