% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/who.R
\name{fetch_who_api_chapter_names}
\alias{fetch_who_api_chapter_names}
\title{Use WHO API to discover chapters}
\usage{
fetch_who_api_chapter_names(ver = "icd10", year = 2016, lang = "en",
  verbose = TRUE)
}
\description{
Of note, the \code{WHO} package does not provide access to classifications, just
WHO summary data.
}
\keyword{internal}
