% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/who.R
\name{fetch_WHO_chapters}
\alias{fetch_WHO_chapters}
\title{Use WHO API to discover chapters}
\usage{
fetch_WHO_chapters(ver = c("icd10", "icd9"), year = 2016,
  lang = "en", verbose = TRUE)
}
\description{
Of note, the \code{WHO} package does not provide access to classifications, just
WHO summary data.
}
\keyword{internal}
