% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bindings.R
\name{get_cim10fr2019}
\alias{get_cim10fr2019}
\title{Semi-localised synonym for \code{\link{get_icd10fr2019}}, with French language column names}
\usage{
get_cim10fr2019()
}
\description{
Classification internationale des maladies.
}
\seealso{
\code{\link{get_icd10fr2019}}
}
\keyword{datasets}
