% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd-cm-ver.R
\name{get_icd10cm_avail_lang}
\alias{get_icd10cm_avail_lang}
\alias{assert_icd10cm_lang_avail}
\title{Get or check available language choices}
\usage{
get_icd10cm_avail_lang(lang)

assert_icd10cm_lang_avail(lang)
}
\description{
Get or check available language choices
}
\keyword{internal}
