% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd-cm-ver.R
\name{get_icd10cm_version}
\alias{get_icd10cm_version}
\title{Get the data for a given version (four-digit year) of ICD-10-CM}
\usage{
get_icd10cm_version(ver = get_icd10cm_active_ver())
}
\arguments{
\item{ver}{Version, currently this is always four-digit year}
}
\description{
Get the data for a given version (four-digit year) of ICD-10-CM
}
\examples{
\dontrun{
get_icd10cm_version("2018")
}
}
