% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/who.R
\name{get_icd10who2016}
\alias{get_icd10who2016}
\alias{get_icd10who2008fr}
\title{Function to get the WHO ICD-10 2016 data}
\usage{
get_icd10who2016(verbose = FALSE)

get_icd10who2008fr()
}
\arguments{
\item{verbose}{logical, default is \code{FALSE}}
}
\description{
This is likely to be removed in the future, as it is designed for the
transition to icd.data version 1.1 . The user may use the active binding
\code{icd10who2016} as if it is a variable. In some situations, it may be
preferable to call this function. E.g., using the active binding when the
cache directory has not been populated may produce messages. Auto-complete in
Rstudio is unfortunately considered still to be interactive, so these
messages may appear prematurely.
}
\details{
This is needed (in addition to icd.data::icd10who2016 active binding),
because lazy data and apparently active bindings are not available until the
package is on the search path. This is difficult or impossible to do in a
CRAN compatible way in the 'icd' package, so this function is just for that
purpose, since we can check whether this function exists and call it without
using ::, whereas this is not possible with lazy data or active bindings.
}
\keyword{datasets}
