% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bindings.R
\name{get_icd9cm2011}
\alias{get_icd9cm2011}
\title{Get the ICD-9-CM data structure}
\usage{
get_icd9cm2011()
}
\description{
This replaces the now deprecated and confusingly named `icd9cm_hierarhcy`
}
\keyword{datasets}
