% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/who.R
\name{get_resource_path}
\alias{get_resource_path}
\alias{set_resource_path}
\title{Get or set the filesystem path used to save downloaded data}
\usage{
get_resource_path()

set_resource_path(path = .default_path, verbose = TRUE)
}
\arguments{
\item{path}{Path to a writable directory}

\item{verbose}{Logical}
}
\description{
Currently this applies only to WHO ICD-10 data, which is downloaded on demand.
The default path is given by \code{file.path(Sys.getenv("HOME"), ".icd.data")}
}
