% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadocs.R
\docType{data}
\name{icd10be2017}
\alias{icd10be2017}
\title{ICD-10-BE 2017}
\description{
ICD-10-BE 2017
}
\seealso{
\code{\link{icd10be2014}} \code{\link{icd10be2014_pc}}
  \code{\link{icd10be2017_pc}}
}
\keyword{datasets}
