% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-file.R
\name{icd10cm_get_flat_file}
\alias{icd10cm_get_flat_file}
\title{Get annual version of ICD-10-CM}
\usage{
icd10cm_get_flat_file(year, verbose = TRUE, ...)
}
\arguments{
\item{year}{four-digit}

\item{...}{passed through, e.g., \code{offline = FALSE}}
}
\description{
Get annual version of ICD-10-CM
}
