% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bindings.R
\docType{data}
\name{icd10cm_latest}
\alias{icd10cm_latest}
\alias{get_icd10cm_latest}
\title{The latest available version of ICD-10-CM in this package}
\usage{
get_icd10cm_latest()
}
\description{
The latest available version of ICD-10-CM in this package
}
\details{
This is an active binding, so is exported explicitly
}
\keyword{datasets}
