% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-icd10.R
\name{icd10cm_parse_all_defined}
\alias{icd10cm_parse_all_defined}
\title{get all ICD-10-CM codes}
\source{
\url{https://www.cms.gov/Medicare/Coding/ICD10} also available from
\url{http://www.cdc.gov/nchs/data/icd/icd10cm/2016/ICD10CM_FY2016_code_descriptions.zip}.
}
\usage{
icd10cm_parse_all_defined(save_data = FALSE, offline = TRUE)
}
\description{
Gets all ICD-10-CM codes from an archive on the CDC web site.
}
\details{
The factor generation uses \code{sort.default} which is locale dependent.
This meant a lot of time debugging a problem when white space was ignored for
sorting on some platforms, but not others.
}
\references{
\href{https://www.cms.gov/Medicare/Coding/ICD10/downloads/icd-10quickrefer.pdf}{CMS ICD-10 Quick Reference}
\href{https://www.cdc.gov/nchs/icd/icd10cm.htm#FY\%202019\%20release\%20of\%20ICD-10-CM}{CDC copy of ICD-10-CM for 2019}
}
\keyword{internal}
