% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd-cm-ver.R
\name{icd10cm_ver_lang_ok}
\alias{icd10cm_ver_lang_ok}
\title{Check whether a given combination of version/year, language, and diagnostic
versus procedure codes is available. Beta.}
\usage{
icd10cm_ver_lang_ok(ver = get_icd10cm_active_ver(),
  lang = get_icd10cm_active_lang(), pc = FALSE)
}
\arguments{
\item{ver}{Character version of ICD-10-CM to use, corresponding to the suffix
of the \code{data.frame} name, e.g. for 2019 ICD-10-CM, use "icd10cm2019" for
Dutch 2014 ICD-CM translations, use 'icd10cm2014_nl"}

\item{lang}{Language string, e.g., "en", "fr", "nl" or others that may be supported in the future.}

\item{pc}{Logical, if \code{TRUE}, will return names of procedure code data
frames. Default is \code{FALSE} which will return the names of the diagnostic
code data frames.}
}
\description{
Check whether a given combination of version/year, language, and diagnostic
versus procedure codes is available. Beta.
}
\examples{
# test active option settings by default
icd10cm_ver_lang_ok()
# French is availabile for 2014 diagnostic codes
icd10cm_ver_lang_ok(2014, "fr")
# Arabic is not available for active year (or any other)
icd10cm_ver_lang_ok(lang = "ar")
}
