% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadocs.R
\docType{data}
\name{icd9cm2011}
\alias{icd9cm2011}
\title{New name for \code{\link{icd9cm_hierarchy}}}
\usage{
icd9cm2011()
}
\description{
New name for \code{\link{icd9cm_hierarchy}}
}
\seealso{
\code{\link{icd9cm_hierarchy}}
}
\keyword{datasets}
\keyword{internal}
