% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadocs.R
\docType{data}
\name{icd9cm_billable}
\alias{icd9cm_billable}
\title{Deprecated ICD-9-CM leaf/billable codes}
\description{
Deprecated ICD-9-CM leaf/billable codes
}
\seealso{
\code{icd9cm2011}, and the also deprecated
  \code{\link{icd9cm_billable}}
}
\keyword{datasets}
