% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R
\name{icd_setup_data_dir}
\alias{icd_setup_data_dir}
\title{Get or set the resource directory for on-demand downloads, and cached data}
\usage{
icd_setup_data_dir(path = NULL,
  interact = getOption("icd.data.interact", FALSE), force = TRUE,
  verbose = TRUE)
}
\arguments{
\item{path}{Path to desired directory}

\item{interact}{Whether to prompt, defaults to \code{interactive()}}

\item{force}{Whether to set the default path of \code{~/.icd.data} regardless
of user interactivity.}

\item{verbose}{single logical value, defaults to \code{FALSE} in most functions.}

\item{must_work}{Single logical, default is \code{FALSE}}

\item{...}{Arguments passed to \code{\link{setup_resource_dir}} if this is
required.}
}
\value{
The path to the resource directory, or \code{NULL} if it could not be
  found.
}
\description{
For getting the path, first the option \code{icd.data.resource} is tried,
otherwise \code{\link{setup_resource_dir}} is called.
}
\details{
For setting, \code{path} is created if it does not exist.
}
\examples{
set_data_dir(td <- tempdir())
get_data_dir()
unlink(td)
try(get_data_dir())
}
