% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cim-10-fr.R
\name{parse_cim_fr}
\alias{parse_cim_fr}
\title{Read the definitions of the French edition of ICD-10}
\usage{
parse_cim_fr(save = FALSE)
}
\arguments{
\item{save}{Logical, default is \code{FALSE}, whether to save results in
package source directory}
}
\description{
The data are encoded, as far as R allows, in UTF-8 for the descriptions, and
without an encoding for the (ASCII) ICD codes themselves.
}
\keyword{internal}
