% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cim-10-fr.R
\name{parse_icd10be2017}
\alias{parse_icd10be2017}
\title{Get French and Dutch translations of ICD-10-CM for Beglian coding}
\usage{
parse_icd10be2017(save_data = FALSE, ...)
}
\arguments{
\item{...}{passed to \code{download_to_data_raw}, e.g., \code{offline = FALSE}.}
}
\description{
There are public domain publications for 2014 and 2017. Not all items are
translated from English. Between 2014 and 2017, ICDLST and ICDLSTR fields 13
and 14 were removed.
}
\references{
\url{https://www.health.belgium.be/fr/sante/organisation-des-soins-de-sante/hopitaux/systemes-denregistrement/icd-10-be}
\url{https://www.health.belgium.be/sites/default/files/uploads/fields/fpshealth_theme_file/fy2017_reflist_icd-10-be.xlsx_last_updatet_28-07-2017_1.xlsx}
\url{https://www.health.belgium.be/fr/fy2014reflisticd-10-bexlsx}
}
\seealso{
\code{link{parse_icd10be2014_be}}
}
\keyword{internal}
