% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/who.R
\name{set_resource_path}
\alias{set_resource_path}
\title{Set the filesystem path used to save downloaded data}
\usage{
set_resource_path(path = getOption("icd.resource.path"),
  verbose = FALSE)
}
\arguments{
\item{path}{Path to a writable directory, default is given by:
\code{file.path(Sys.getenv("HOME"), ".icd.data")}}

\item{verbose}{Logical}
}
\value{
The final path, which may be temporary directory if user denied
creation of home directory hidden directory.
}
\description{
If the directory exists, the directory is used by \code{icd.data}. If the
directory does not exist, and R is in \code{interactive()} mode, then the user is
asked whether a directory may be created. The default path of this directory
is given by \code{file.path(Sys.getenv("HOME"), ".icd.data")}, which is
\code{~/.icd.data} on Linux, MacOS, UNIX. Something like
\code{options(icd.data.resource = "/tmp/icd.data.resource")} may be used to change
the default directory attempted, or to avoid a prompt by using that directory
automatically. This may be set in \code{.Rprofile}. The simplest way is just to
let \code{icd.data} use the default directory.
}
\section{Non-interactive mode}{
 If R is running non-interactively, e.g. CRAN
testing, Travis, package building, or in an automated user script, no
directories are created and a temporary directory is created. The
session-wide R option \code{icd.data.resource} is set to the temporary
directory's path.
}

\seealso{
\link{get_resource_path}
}
