% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R, R/resource.R
\name{setup_icd_data}
\alias{setup_icd_data}
\alias{icd_data_dir}
\title{Set up the data download cache, give permission to download data}
\usage{
setup_icd_data(path = .icd_data_default)

icd_data_dir()
}
\arguments{
\item{path}{Path to a directory where cached online raw and parsed data will be cached. It will be created if it doesn't exist.}
}
\value{
The path to the resource directory, or \code{NULL} if it could not be
  found.

Invisiblly returns the data path which was set, or NULL if not done.
}
\description{
This must be called by the user, as prompted on package attach with \code{library(icd.data)}. \code{icd.data} is a dependency (not an import) of \code{icd}, so that \code{icd} can function more smoothly, avoiding prompting during commands, although this should still be possible, and will happen if the user initially declines permission to download and cache data.
}
\section{Functions}{
\itemize{
\item \code{icd_data_dir}: Return the currently active data directory. If missing, it will return \code{NULL} and, depending on \code{getOption("icd.data.absent_action")}, will stop, give a message, or do nothing.
}}

\examples{
\dontrun{
setup_icd_data()
setup_icd_data("/var/cache/icd.data")
setup_icd_data(path = ".local/icd.data")
}
# or use an unexported function to set or reset:
icd.data:::.set_data_dir(td <- tempdir())
# what is the currently set directory?
icd_data_dir()
unlink(td)
try(icd_data_dir())
}
