% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R
\name{setup_resource_dir}
\alias{setup_resource_dir}
\alias{set_resource_dir}
\alias{get_resource_dir}
\title{This directory is created only at user's permission. Default is \code{~/.icd.data}.
Get or set the resource directory for on-demand downloads, and cached data}
\usage{
setup_resource_dir(path = NULL, interact = interactive(),
  force = TRUE, verbose = TRUE)

set_resource_dir(path)

get_resource_dir(must_work = FALSE, ...)
}
\arguments{
\item{path}{Path to desired directory}

\item{interact}{Whether to prompt, defaults to \code{interactive()}}

\item{force}{Whether to set the default path of \code{~/.icd.data} regardless
of user interactivity.}

\item{verbose}{single logical value, defaults to \code{FALSE} in most functions.}

\item{must_work}{Single logical, default is \code{FALSE}}

\item{...}{Arguments passed to \code{\link{setup_resource_dir}} if this is
required.}
}
\value{
The path to the resource directory, or \code{NULL} if it could not be
found.
}
\description{
For getting the path, first the option \code{icd.data.resource} is tried,
otherwise \code{\link{setup_resource_dir}} is called.
}
\details{
For setting, \code{path} is created if it does not exist.
}
\examples{
set_resource_dir(td <- tempdir())
get_resource_dir()
unlink(td)
try(get_resource_dir())
}
