% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-test.R
\name{skip_missing_icd10who}
\alias{skip_missing_icd10who}
\title{Skip subsequent tests that depend on WHO ICD-10 data if not available}
\usage{
skip_missing_icd10who(ver = "2016")
}
\arguments{
\item{ver}{Version of WHO ICD-10 to use, currently a four-digit year}
}
\description{
Will only skip if the data is absent
}
