% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-test.R
\name{skip_missing_icd10who2016}
\alias{skip_missing_icd10who2016}
\title{Skip subsequent tests that depend on WHO ICD-10 data if not available}
\usage{
skip_missing_icd10who2016()
}
\description{
Will only skip if the data is absent
}
