% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd-cm-ver.R
\name{.get_icd10cm_latest}
\alias{.get_icd10cm_latest}
\title{Get the data for the latest ICD-10-CM version in this package.}
\usage{
.get_icd10cm_latest()
}
\description{
Usually, the \code{icd.data} package should be attached, and therefore appear on
the search list, using \code{library(icd.data)} or \code{require(icd.data)}. If it is
not attached, the active binding \code{icd10cm_latest} cannot find the packages
own data! This function is a possible work-around to get the data without
having to attach the package.
}
\examples{
\dontrun{
# if icd.data not attached:
get_icd10cm_latest()
icd.data::icd10cm_latest
# preferred:
library(icd.data)
head(icd10cm_latest)
}
}
\keyword{datasets}
\keyword{internal}
