% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cim-10-fr.R
\name{.parse_icd10be2014}
\alias{.parse_icd10be2014}
\title{Get 2014 French and Dutch translations of ICD-10-CM for Beglian coding}
\usage{
.parse_icd10be2014(save_data = FALSE, ...)
}
\arguments{
\item{...}{passed to \code{.download_to_data_raw}, e.g., \code{offline = FALSE}.}
}
\description{
Get 2014 French and Dutch translations of ICD-10-CM for Beglian coding
}
\references{
\url{https://www.health.belgium.be/fr/sante/organisation-des-soins-de-sante/hopitaux/systemes-denregistrement/icd-10-be}
\url{https://www.health.belgium.be/fr/fy2014reflisticd-10-bexlsx}
}
\seealso{
\code{link{parse_icd10be2014_be}}
}
\keyword{internal}
