% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/who.R
\name{fetch_icd10who2016}
\alias{fetch_icd10who2016}
\alias{fetch_icd10who2008_fr}
\title{Fetch the WHO data from online source}
\usage{
fetch_icd10who2016(save_data = TRUE, ...)

fetch_icd10who2008_fr(save_data = FALSE, ...)
}
\arguments{
\item{save_data}{Logical, defaults to \code{TRUE}}

\item{...}{Arguments passed to internal functions}
}
\description{
This will download the latest ICD-10 codes and descriptions from the WHO, and
save the results in a directory given by \code{icd.data:::get_resource_path()}.
This defaults to a subdirectory \code{.icd.data} of the home directory. This is
necessary because it is not permitted to write data back to the installed
package location (and this may not be allowed on a multi-user system,
anyway).
}
