% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd-cm-ver.R
\name{get_icd10cm_latest}
\alias{get_icd10cm_latest}
\title{Get the data for the latest ICD-10-CM version in this package.}
\usage{
get_icd10cm_latest()

get_icd10cm_latest()
}
\description{
Language handling is incomplete, as there is only data for 2014 for French
and Dutch. These functions are beta status, and will likely need updating as
more national ICD versions, years, sub-versions, and languages are added.
}
\details{
Usually, the \code{icd.data} package should be attached, and therefore appear on
the search list, using \code{library(icd.data)} or \code{require(icd.data)}. If it is
not attached, the active binding \code{icd10cm_latest} cannot find the packages
own data! This function is a possible work-around to get the data without
having to attach the package.
}
\examples{
\dontrun{
# if icd.data not attached:
get_icd10cm_latest()
# preferred:
library(icd.data)
head(icd10cm_latest)

}
# For the latest data, regardless of active settings:
identical(icd10cm_latest, icd10cm_active)
identical(icd10cm_latest, icd10cm2019)
# note a change between 2018 and 2019
icd10cm_latest[icd10cm_latest$code == "C4311", "billable"]
icd10cm2018[icd10cm2018$code == "C4311", "billable"]
"C43111" \%in\% icd10cm2018$code
icd10cm2019[icd10cm2019$code == "C43111", ]
}
\keyword{datasets}
\keyword{internal}
