% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/who-internal.R
\name{get_resource_path}
\alias{get_resource_path}
\title{Get the filesystem path used to save downloaded data}
\usage{
get_resource_path(set_if_missing = FALSE, must_work = TRUE,
  verbose = FALSE)
}
\arguments{
\item{set_if_missing}{Logical, if \code{TRUE}, \link{set_resource_path} is called}

\item{must_work}{Logical, if \code{TRUE}, the default, failure to get a valid path
will stop execution.}

\item{verbose}{Logical, default \code{FALSE}}
}
\description{
Currently this is only needed for WHO ICD-10 data, which is downloaded on
demand. The default path attempted is \code{file.path(Sys.getenv("HOME"), ".icd.data")}. This may be overridden by setting the option
\code{icd.data.resource} to the path of your choice. E.g.,
\code{set_resource_path('/tmp/icd.data.resource')} or by setting the option itself
using \code{options(icd.data.resource = '/tmp/icd.data.resource')}.
}
\seealso{
\link{set_resource_path}
}
