% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd.data-package.R
\docType{package}
\name{icd.data}
\alias{icd.data}
\alias{_PACKAGE}
\alias{icd.data-package}
\title{icd.data: International Classification of Diseases (ICD) Data}
\description{
ICD-9 and ICD-10 definitions from the World Health
    Organization, the United States Center for Medicare and Medicaid
    Services (CMS), France, and Belgium are made available by this
    package. There are diagnostic and procedure codes, and lists of the
    chapter and sub-chapter headings and the ranges of ICD codes they
    encompass. There are also two sets of sample patient data with ICD-9
    and ICD-10 codes representing real patients. Much of the data is
    downloaded and parsed on demand. This can all be done at once, or
    piecemeal as data is used.  These data are used by the 'icd' package,
    which finds comorbidities and has tools for working with ICD codes.
}
\author{
Jack O. Wasey \email{jack@jackwasey.com}
}
\concept{ICD-9 ICD-10 ICD}
\keyword{datasets}
\keyword{misc}
\keyword{utilities}
