% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign.R
\docType{package}
\name{icd.data}
\alias{icd.data}
\alias{_PACKAGE}
\alias{icd.data-package}
\alias{icd.data-package}
\title{icd.data: International Classifcation of Diseases (ICD) Data}
\description{
Data from the United States Center for Medicare and Medicaid
Services (CMS) is included in this package. There are ICD-9 and ICD-10
diagnostic and procedure codes, and lists of the chapter and sub-chapter
headings and the ranges of ICD codes they encompass. There are also two
sample datasets. These data are used by the 'icd' package for finding
comorbidities.
}
\author{
Jack O. Wasey \email{jack@jackwasey.com}
}
\concept{
ICD-9 ICD-10 ICD
}
