% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadocs.R
\docType{data}
\name{icd10cm_active}
\alias{icd10cm_active}
\title{The currently active version of ICD-10-CM in this package.}
\description{
The default is the latest version, but it may be set using \code{set_icd10cm_ver()}
}
\details{
This is an active binding.
}
\examples{
\dontrun{
set_icd10cm_active_ver("2016")
}
}
\keyword{datasets}
