% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadocs.R
\docType{data}
\name{icd10cm_latest}
\alias{icd10cm_latest}
\title{The latest version of ICD-10-CM available in this package}
\description{
The latest version of ICD-10-CM available in this package
}
\details{
This is an active binding to the latest data.
}
\keyword{sysdata}
