% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadocs.R
\docType{data}
\name{icd10fr2018}
\alias{icd10fr2018}
\title{Définitions CIM-10-FR}
\source{
\url{https://www.atih.sante.fr/cim-10-fr-2018-usage-pmsi}
}
\description{
La Classification internationale statistique des maladies (CIM), version 10,
edition francaise. (The International Classification of Diseases (ICD),
version 10, French edition.) Comme la version américaine, l'édition française
a beaucoup de changéements par rapport à l'édition de l'OMS. Juste l'année
2018 est présentée pour le moment.
}
\keyword{datasets}
