% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadocs.R
\docType{data}
\name{icd9cm_billable}
\alias{icd9cm_billable}
\title{Deprecated name for the ICD-9-CM leaf nodes}
\format{data frames with columns \code{code}, \code{short_desc}, and
  \code{long_desc}.}
\source{
\url{http://www.cms.gov/Medicare/Coding/ICD9ProviderDiagnosticCodes/codes.html}
}
\description{
This now just contains the final version (2011, version number 32). Use
\code{\link{icd9cm_leaf_v32}} instead.
}
\seealso{
\code{\link{icd9cm_leaf_v32}}
}
\keyword{datasets}
