% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd-cm-ver.R
\name{set_icd10cm_active_ver}
\alias{set_icd10cm_active_ver}
\alias{get_icd10cm_active_ver}
\alias{set_icd10cm_active_lang}
\alias{get_icd10cm_active_lang}
\title{Get or set the annual version or language of ICD-10-CM to use}
\usage{
set_icd10cm_active_ver(ver)

get_icd10cm_active_ver()

set_icd10cm_active_lang(lang)

get_icd10cm_active_lang()
}
\arguments{
\item{ver}{Character version of ICD-10-CM to use, corresponding to the suffix
of the \code{data.frame} name, e.g. for 2019 ICD-10-CM, use "icd10cm2019" for
Dutch 2014 ICD-CM translations, use 'icd10cm2014_nl"}

\item{lang}{Language string, e.g., "en", "fr", "nl" or others that may be supported in the future.}
}
\description{
Get or set the annual version or language of ICD-10-CM to use
}
\examples{
  set_icd10cm_active_ver(2019)
}
